Move VirtualFloor.cpp to the paint folder and fix Xcode project
