+++ src/openrct2/paint/VirtualFloor.h
